requirements.txt
altair==5.2.0
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==23.1.0
blinker==1.9.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
Django==5.1.5
Flask==3.1.1
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.6
importlib-metadata==6.8.0
itsdangerous==2.2.0
Jinja2==3.1.2
jiter==0.8.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
numpy==1.26.2
openai==1.58.1
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
protobuf==4.25.1
pyarrow==14.0.1
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.1.0
pytz==2023.3.post1
redis==6.1.0
referencing==0.31.1
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
spotipy==2.25.1
sqlparse==0.5.3
streamlit==1.28.2
sympy==1.13.3
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
Werkzeug==3.1.3
zipp==3.17.0
